Drawing grid by using a threejs helper.
I needed to be able to write to the depth buffer in order to draw the
cartesian plane with shaders. Since this extension is not working on my
machine, I'll just go forward with THREE Js grid helper, which also has
the advantage of working well in ill posed cases.