Proceeding with mesh rendering.
Seems like FBX isn't going to work out since there are no readily,
updated libraries that export from FBX to three.js. The one that comes
with threejs exports to a Scene format, and SceneLoader is both
deprecated and doesn't call the callback routine (I tested it with
dude.fbx).